+++ market/src/main/java/com/challenge/market/domain/member/repository/MemberRepository.java
@@ -1,0 +1,13 @@
+package com.challenge.market.domain.member.repository;
+
+import com.challenge.market.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member,Long> {
+    Optional<Member> findByName(String name);
+}
