+++ market/src/main/java/com/challenge/market/domain/product/constant/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.challenge.market.domain.product.constant;
+
+public enum ProductStatus {
+    SALES, COMPLETION , RESERVATION
+}
