+++ market/src/main/java/com/challenge/market/domain/product/repository/ProductRepository.java
@@ -1,0 +1,11 @@
+package com.challenge.market.domain.product.repository;
+
+import com.challenge.market.domain.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+
+}
