--- market/src/main/java/com/challenge/market/member/dto/SignUpRequest.java
@@ -1,25 +1,0 @@
-package com.challenge.market.member.dto;
-
-import com.challenge.market.member.domain.Member;
-import jakarta.validation.constraints.NotEmpty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignUpRequest {
-
-    @NotEmpty
-    private String name;
-
-    @NotEmpty
-    private String pw;
-
-
-
-
-}
