--- market/src/main/java/com/challenge/market/member/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.challenge.market.member.repository;
-
-import com.challenge.market.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member,Long> {
-}
