--- market/src/main/java/com/challenge/market/product/advice/ProductControllerAdvice.java
@@ -1,10 +1,0 @@
-package com.challenge.market.product.advice;
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-@RestControllerAdvice
-public class ProductControllerAdvice extends ResponseEntityExceptionHandler {
-
-
-}
