--- market/src/main/java/com/challenge/market/product/domain/constant/ProductStatus.java
@@ -1,5 +1,0 @@
-package com.challenge.market.product.domain.constant;
-
-public enum ProductStatus {
-    SALES, COMPLETION , RESERVATION
-}
