--- market/src/main/java/com/challenge/market/product/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.challenge.market.product.repository;
-
-import com.challenge.market.product.domain.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-
-}
