+++ market/src/main/java/com/challenge/market/util/SessionUtil.java
@@ -1,0 +1,6 @@
+package com.challenge.market.util;
+
+public interface SessionUtil {
+
+    public static final String LOGIN_ID = "loginId";
+}
