+++ market/src/main/java/com/challenge/market/web/error/ErrorResponse.java
@@ -1,0 +1,11 @@
+package com.challenge.market.web.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class ErrorResponse {
+    private String status;
+    private String message;
+}
