+++ market/src/main/java/com/challenge/market/web/member/dto/SignInMember.java
@@ -1,0 +1,17 @@
+package com.challenge.market.web.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class SignInMember {
+
+    private String memberId;
+    private String memberName;
+
+}
