+++ market/src/main/java/com/challenge/market/web/member/dto/SignUpRequest.java
@@ -1,0 +1,24 @@
+package com.challenge.market.web.member.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignUpRequest {
+
+    @NotEmpty
+    private String name;
+
+    @NotEmpty
+    private String pw;
+
+
+
+
+}
