+++ market/src/main/java/com/challenge/market/web/product/handler/ProductControllerAdvice.java
@@ -1,0 +1,10 @@
+package com.challenge.market.web.product.handler;
+
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+@RestControllerAdvice
+public class ProductControllerAdvice extends ResponseEntityExceptionHandler {
+
+
+}
